requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
PyJWT==2.7.0
